pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "etl"
version = "0.1.0"
description = "ETL pipeline for processing Chinese-language documents"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "YiLu", email = "yilu@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "pyyaml>=5.4.0",
    "python-dotenv>=0.19.0",
    "loguru>=0.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "isort>=5.10.0",
    "mypy>=0.950",
]
notebooks = [
    "jupyter>=1.0.0",
    "notebook>=6.4.0",
    "jupyterlab>=3.0.0",
    "ipython>=7.30.0",
]

[tool.setuptools]
packages = ["etl"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
etl = ["config/*.yaml", "config/*.yml"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=src/etl --cov-report=html --cov-report=term-missing"
